+++ NEWS.md
@@ -6,6 +6,14 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.0.3
+
+- Features
+  - Force update all failed imports admin task.
+- Fixes
+  - Fix moving folders to subfolder orphans folders bug.
+  - Fix id does not exist redirect.
+
 ## v1.0.2
 
 - Features